Explode the code so that the backend is explicit.
This change is required so that adding WebGL / OpenGL ES won’t make us
go bad by changing all the existing code. The implication is going to
be, in the first place, code duplication. As soon as we will hit two /
three backends, we should be able to find a suitable set of associated
types in GraphicsContext to allow to create e.g. a buffer agnostic of
the backend.